Source package version from setup.py.
This avoids a tricky chicken-or-egg problem:
as soon as shimmingtoolbox/__init__.py starts loading up
submodules, which are going to load up libraries,
installation will fail because setup.py was doing

> import shimmingtoolbox

at the top, just to get at

> shimmingtoolbox.__version__

and of course when setup.py is run, none of the
libraries we need will be installed yet.

This is mentioned on https://packaging.python.org/guides/single-sourcing-package-version/

> Warning
>
> Although this technique is common, beware that it will fail if sample/__init__.py
> imports packages from install_requires dependencies, which will very likely not
> be installed yet when setup.py is run.

and after looking through that page, I decided that using importlib.metadata
is the future-proofed option.

(Editorializing: I really wish PyPA would stick to the Zen of Python's admonishment
to "There should be one obvious way to do it."; but the first thing this page says is
"There are many techniques" and the best is buried as #5)