Upgrade dcm2niix to latest.
dcm2bids has only ever been tested with the latest version,
and dcm2niix is still shaking out its API bugs, so the only
reliable thing to do is to demand everyone stay up to date.

This is okay almost everywhere except on Debian which is always
slow to upgrade stuff.