[fix][download_data] Fix CLI help message
**Why this change was necessary**
The desired behaviour is to automatically generate a list of data
available using the URL dictionary.

**What this change does**
Modifies the format of the URL dictionary to include the dataset
description. Then, the URL dictionary is walked to generate a
multiline string containing the available dataset names and
descriptions. This string description is then used in the `help` kw
arg in the `click` command decorator for `main()`.

Note that I'm not sure what `prelude` does, so I have a TODO left in
the source code until I have more information.

**Any side-effects?**
The spacing of the output is wonky as a result of the double
newline (\n) used. I tried using a single newline, but click seems to
get rid of those, which is *not* the behaviour of attempting to print
an f-string containing single newline chars:

```
print(f"hello\nhi")
```

This forced me to use double newline chars for each dataset option.

Furthermore, any attempt to indent each option doesn't seem to work
either, so I suspect that `click` does some kind of whitespace
trimming.

The resulting output of running the file with the `--help` flag:

```
$ python download_data.py --help
Usage: download_data.py [OPTIONS] DATA

  Download data from the internet. The available datasets are:

  - testing_data: Light-weighted dataset for testing purpose.

  - prelude: Binary for prelude software

Options:
  --verbose      Be more verbose.
  --output TEXT  Output folder.
  --help         Show this message and exit.
```

**Additional context/notes/links**

Related to #91 - Address TODOs in cli/download_data.py