[chore][meta] Satisfy GitHub community profile checklist
**Why this change was necessary**
We want to make our project as friendly as possible for outsider
contributions. These may be in the form of both issue and PR
submissions. GitHub provides a checklist of things to include a
repository to encourage good contributions.

This is particularly important for prepping our project for
Hacktoberfest. Additionally, this invested effort will also speed up
onboarding of new lab members and collaborators.

**What this change does**
Adds issue templates for bug reports, enhancement requests for
existing features, and feature requests for new features. Adds a PR
template consistent with NeuroPoly contribution guidelines. Removes PR
guidelines from RTD docs to ensure the PR and issue templates are the
single source of truth.

Adds a code of conduct (The Contributor Covenant) that all
contributors must agree to follow so we can have a safe and productive
environment for all.

**Any side-effects?**
None

**Additional context/notes/links**

Resolves #105 - Complete community profile checklist