[test][conftest] Download test_data before running test suite
**Why this change was necessary**
We want to ensure that test data has been pulled into the test
environment before running tests.

**What this change does**
Executes the download_data() CLI inside a pytest hook that runs first
when the test suite session begins.

**Any side-effects?**
Creates a new temp directory and populates it with test data.

Should this temp directory be removed after the testing run is
complete? Can we expect that the testing environment will be connected
to the internet?

**Additional context/notes/links**

Resolves #118 - Make sure the environment is properly setup and
binaries are present before running pytest

May want to refactor with an @setup decorator:
https://stackoverflow.com/a/12600154/6310633.