[test][prelude,dcm2niix] Add pytest markers for soft dependencies
**Why this change was necessary**
We want to be able to conditionally run some tests for interfacing
shimming-toolbox with its soft dependencies.

Currently, `prelude` is the only tool with its functionality, but in
the future, users may wish to use a different tool covering the same
functionality. Therefore, tests requiring interfacing with `prelude`
don't need to be run if `prelude` is not installed, and this case must
be covered by the shimming-toolbox testing suite.

`dcm2niix` is only needed by the `dcm2bids` library, which itself is
only needed by users who don't already have their files in the BIDS
format. If they already have BIDS files, they don't need to install
`dcm2niix` and therefore the shimming-toolbox testing suite must cover
the case where dcm2niix is not installed.

**What this change does**
This change uses the pytest feature allowing the marking of test
functions with specific markers[1]. Marked test functions only run
when the corresponding marker is passed as an argument to the `-m`
option to pytest. This change also updates the accompanying usage
instructions for this feature.

**Any side-effects?**
The minimal `$ pytest` command still runs all test functions. You must
specifically opt out of tests depending on the soft dependencies. See
accompanying documentation for details.

**Rejected implementation options**
pytest provides the functionality for custom markers specified at
runtime[3]. This option was rejected as each custom environment is
only defined inline in the decorator, which could lead to divergence
of decorators during an incomplete refactor. The chosen implementation
allows configuration of markers in a single source of truth,
simplifying the refactor and addition of new markers.

The desired functionality could also be implemented using the
`pytest.mark.skipif` pattern and hooking that to an automated dependency
check. However, this creates implicit rather than explicit
behaviour and could result in duplicated condition definitions in each
decorator.

pytest also provides functionality for running only tests that match a
provided keyword expression[4], but this approach was rejected as it
relies on consistent naming conventions across existing tests
depending on either of the soft dependencies. Using this approach
would require a significant refactor of function and accompanying test
function names.

**Additional context/notes/links**

You can also mark whole modules with a `pytest.mark` decorator[2].

Note that the markers are defined in a SSOT in `pytest.ini`. New
markers can be defined in this file, following the existing
formatting. The decorators for test functions are in the format
`@pytest.mark.markername`.

See also:
1: https://docs.pytest.org/en/latest/example/markers.html
2:
https://docs.pytest.org/en/latest/example/markers.html#marking-whole-classes-or-modules
3:
https://docs.pytest.org/en/latest/example/markers.html#custom-marker-and-command-line-option-to-control-test-runs
4: https://docs.pytest.org/en/latest/usage.html#specifying-tests-selecting-tests

Resolves #118 - Make sure the environment is properly setup and binaries are present before running pytest