[test][conftest] Test existence of soft dependencies conditionally
**Why this change was necessary**
We only need to check the existence of soft dependencies when we're
testing for those dependencies. Therefore, these tests should be
decorated with the corresponding mark.

**What this change does**
Adds test functions in conftest to ensure that the soft dependencies
are installed and available to shimming-toolbox.

**Any side-effects?**
None.

**Additional context/notes/links**

Note that markers cannot be easily added to fixtures. I used a
workaround documented in pytest's issues[1].

Fixtures cannot be passed to a class signature, and it would
duplicate a lot of code to pass to every test class methods, so
instead we can use the `@pytest.mark.usefixtures()`
decorator[2]. Unfortunately, this cannot be used for multiple
functions within a test file unless it contains a test class[2.Warning].

1: https://github.com/pytest-dev/pytest/issues/1368#issuecomment-466339463
2: https://docs.pytest.org/en/stable/fixture.html#use-fixtures-in-classes-and-modules-with-usefixtures

Resolves #118 - Make sure the environment is properly setup and binaries are present before running pytest