Ensure test environment is set up with test data and soft dependencies before running tests (#136)
* [test][conftest] Download test_data before running test suite

**Why this change was necessary**
We want to ensure that test data has been pulled into the test
environment before running tests.

**What this change does**
Executes the download_data() CLI inside a pytest hook that runs first
when the test suite session begins.

**Any side-effects?**
Creates a new temp directory and populates it with test data.

Should this temp directory be removed after the testing run is
complete? Can we expect that the testing environment will be connected
to the internet?

**Additional context/notes/links**

Resolves #118 - Make sure the environment is properly setup and
binaries are present before running pytest

* [refactor][conftest] Use __dir_testing__ pkg variable

**Why this change was necessary**
The package already has a global available for the testing dir
location and dirname.

**What this change does**
See title.

**Any side-effects?**
None.

**Additional context/notes/links**

Related to https://github.com/shimming-toolbox/shimming-toolbox/pull/136#issuecomment-705301251

* [style][tests] Reformat test scripts

* [test][prelude,dcm2niix] Add pytest markers for soft dependencies

**Why this change was necessary**
We want to be able to conditionally run some tests for interfacing
shimming-toolbox with its soft dependencies.

Currently, `prelude` is the only tool with its functionality, but in
the future, users may wish to use a different tool covering the same
functionality. Therefore, tests requiring interfacing with `prelude`
don't need to be run if `prelude` is not installed, and this case must
be covered by the shimming-toolbox testing suite.

`dcm2niix` is only needed by the `dcm2bids` library, which itself is
only needed by users who don't already have their files in the BIDS
format. If they already have BIDS files, they don't need to install
`dcm2niix` and therefore the shimming-toolbox testing suite must cover
the case where dcm2niix is not installed.

**What this change does**
This change uses the pytest feature allowing the marking of test
functions with specific markers[1]. Marked test functions only run
when the corresponding marker is passed as an argument to the `-m`
option to pytest. This change also updates the accompanying usage
instructions for this feature.

**Any side-effects?**
The minimal `$ pytest` command still runs all test functions. You must
specifically opt out of tests depending on the soft dependencies. See
accompanying documentation for details.

**Rejected implementation options**
pytest provides the functionality for custom markers specified at
runtime[3]. This option was rejected as each custom environment is
only defined inline in the decorator, which could lead to divergence
of decorators during an incomplete refactor. The chosen implementation
allows configuration of markers in a single source of truth,
simplifying the refactor and addition of new markers.

The desired functionality could also be implemented using the
`pytest.mark.skipif` pattern and hooking that to an automated dependency
check. However, this creates implicit rather than explicit
behaviour and could result in duplicated condition definitions in each
decorator.

pytest also provides functionality for running only tests that match a
provided keyword expression[4], but this approach was rejected as it
relies on consistent naming conventions across existing tests
depending on either of the soft dependencies. Using this approach
would require a significant refactor of function and accompanying test
function names.

**Additional context/notes/links**

You can also mark whole modules with a `pytest.mark` decorator[2].

Note that the markers are defined in a SSOT in `pytest.ini`. New
markers can be defined in this file, following the existing
formatting. The decorators for test functions are in the format
`@pytest.mark.markername`.

See also:
1: https://docs.pytest.org/en/latest/example/markers.html
2:
https://docs.pytest.org/en/latest/example/markers.html#marking-whole-classes-or-modules
3:
https://docs.pytest.org/en/latest/example/markers.html#custom-marker-and-command-line-option-to-control-test-runs
4: https://docs.pytest.org/en/latest/usage.html#specifying-tests-selecting-tests

Resolves #118 - Make sure the environment is properly setup and binaries are present before running pytest

* [test][conftest] Test existence of soft dependencies conditionally

**Why this change was necessary**
We only need to check the existence of soft dependencies when we're
testing for those dependencies. Therefore, these tests should be
decorated with the corresponding mark.

**What this change does**
Adds test functions in conftest to ensure that the soft dependencies
are installed and available to shimming-toolbox.

**Any side-effects?**
None.

**Additional context/notes/links**

Note that markers cannot be easily added to fixtures. I used a
workaround documented in pytest's issues[1].

Fixtures cannot be passed to a class signature, and it would
duplicate a lot of code to pass to every test class methods, so
instead we can use the `@pytest.mark.usefixtures()`
decorator[2]. Unfortunately, this cannot be used for multiple
functions within a test file unless it contains a test class[2.Warning].

1: https://github.com/pytest-dev/pytest/issues/1368#issuecomment-466339463
2: https://docs.pytest.org/en/stable/fixture.html#use-fixtures-in-classes-and-modules-with-usefixtures

Resolves #118 - Make sure the environment is properly setup and binaries are present before running pytest

* [chore][conftest] Clean up WIP changes

**Why this change was necessary**
I left some WIP code and comments.

**What this change does**
See title.

**Any side-effects?**
New output message upon download completion.

**Additional context/notes/links**

Resolves #118 - Make sure the environment is properly setup and binaries are present before running pytest

* [tests][conftest] Update header

* Revert "[style][tests] Reformat test scripts"

This reverts commit 1b5cc4dd16986b90fd2f37ceda3747da93a2a7b3.

* [fix][test_prelude] Restore missing import

* [refactor][conftest] Clarify variable names to reflect purpose

**Why this change was necessary**
The previous names of variables for paths did not accurately reflect
their purposes.

**What this change does**
See title.

**Any side-effects?**
None

**Additional context/notes/links**

Resolves https://github.com/shimming-toolbox/shimming-toolbox/pull/136#issuecomment-706474241
Related to #118

* [test][Travis] Remove test data download from Travis config

**Why this change was necessary**
Since the testing suite already downloads the test data, we can remove
this functionality from the Travis build setup.

**What this change does**
See title.

**Any side-effects?**
Removes redundancy of test data download.

**Additional context/notes/links**

Resolves https://github.com/shimming-toolbox/shimming-toolbox/pull/136#issuecomment-706474241
Related to #118