Update shimming-toolbox package install (#146)
* [fix][Travis] Update shimming-toolbox package install

**Why this change was necessary**
The install command referenced a deprecated extra for the package
install directive. This testing extra was already added to the default
installation.

**What this change does**
Removes the old extra as well as the -e flag since installing the
project in editable mode is not necessary in the Travis environment.

**Any side-effects?**
1 less warning in the build logs!

**Additional context/notes/links**

Resolves #145 - Update Travis build with correct pip install for shimming-toolbox
Related to #97

* [fix][installation] Add __init__.py for CLI module

**Why this change was necessary**
Previously the CLI module did not have an `__init__.py`, meaning that
it was possible to import it, and that the CLI entry points of
shimming-toolbox were invalid, throwing `ModuleNotFoundError`. This bug was previously unnoticed
because installing the package with `pip install -e` seems to ignore
missing `__init__.py` files.

**What this change does**
See title. Adding `cli/__init__.py` allows the entry point CLI tools
to work properly without throwing `ModuleNotFoundError`.

**Any side-effects?**
None

**Additional context/notes/links**

Resolves #145 - Update Travis build with correct pip install for shimming-toolbox
Related to https://github.com/shimming-toolbox/shimming-toolbox/pull/146#issuecomment-706638231

* [chore][setup.py] Remove old subpackages to exclude

**Why this change was necessary**
"contrib" and "tests" are no longer subpackages to be excluded by
`setup.py`.

**What this change does**
See title.

**Any side-effects?**
None.

**Additional context/notes/links**

Resolves #145 - Update Travis build with correct pip install for shimming-toolbox