Add CLI to check 3rd-party binary installation
Resolves #161

**Why this change was necessary**
It's more useful to the user to have a separate script that they can
run to check their installation without running the visually
overwhelming testing suite.

**What this change does**
Creates 2 helper functions using the GNU `which` tool to verify that
the `prelude` and `dcm2niic` binaries are both installed and available
in the PATH. Since these helper functions only return the resulting
exit code of the `which` calls, a wrapper function is exposed as a
CLI, with informative text for a human user.

**Any side-effects?**
Note that this duplicates code in conftest.py, which will be
refactored in a subsequent commit to rely on the helper functions
added in this commit.