Add helper function to check dcm2niix version
Resolves #161

**Why this change was necessary**
Part of the issue spec is to also check the dependency version if it's
installed.

**What this change does**
Adds function that captures output of `dcm2niix --version` and is
called if `dcm2niix` has been verified to be installed and in the
PATH.

Note that `dcm2niix --version` returns a nonzero exit code. This
helper's approach will be invalid if this behaviour in dcm2niix is
fixed.

**Any side-effects?**
None.