Add helper function to check prelude version
Resolves #161

**Why this change was necessary**
Part of the issue spec is to also check the dependency version if it's
installed.

**What this change does**
Adds function that captures output of `prelude --help` and is
called if `prelude` has been verified to be installed and in the PATH.

**Any side-effects?**
Note that `prelude` displays version information as output from the
`--help` flag, but errors out. Therefore, the version must be parsed
from a general help message that's captured from stderr rather than
stdout. This approach will be invalid if/when FSL fixes this
behaviour.