Add CLI to dump environment and package info
Resolves #161

**Why this change was necessary**
It's useful to the user to have a CLI tool to dump environment info
into stdout to help ST contributors with debugging.

**What this change does**
Adds 2 helper functions to get environment information including OS
and CPU details, as well as resource usage.

**Any side-effects?**
New dependency is added, so the shimmingtoolbox package will have to
be reinstalled locally.

**Additional context/notes/links**
https://github.com/shimming-toolbox/shimming-toolbox/issues/161#issuecomment-725028653