Check installation and versions of 3rd party binaries (#176)
* Add CLI to check 3rd-party binary installation

Resolves #161

**Why this change was necessary**
It's more useful to the user to have a separate script that they can
run to check their installation without running the visually
overwhelming testing suite.

**What this change does**
Creates 2 helper functions using the GNU `which` tool to verify that
the `prelude` and `dcm2niic` binaries are both installed and available
in the PATH. Since these helper functions only return the resulting
exit code of the `which` calls, a wrapper function is exposed as a
CLI, with informative text for a human user.

**Any side-effects?**
Note that this duplicates code in conftest.py, which will be
refactored in a subsequent commit to rely on the helper functions
added in this commit.

* Refactor conftest to use check_env helpers

Related to #161

**Why this change was necessary**
conftest.py had some functionality duplicated in check_env.py, so the
duplicated code was factored out.

**What this change does**
Refactors conftest dependency checks to rely on imported check_env.py.

**Any side-effects?**
conftest now depends on the check_env helper functions for verifying
the prelude and dcm2niix installations.

* Add helper function to check dcm2niix version

Resolves #161

**Why this change was necessary**
Part of the issue spec is to also check the dependency version if it's
installed.

**What this change does**
Adds function that captures output of `dcm2niix --version` and is
called if `dcm2niix` has been verified to be installed and in the
PATH.

Note that `dcm2niix --version` returns a nonzero exit code. This
helper's approach will be invalid if this behaviour in dcm2niix is
fixed.

**Any side-effects?**
None.

* Add helper function to check prelude version

Resolves #161

**Why this change was necessary**
Part of the issue spec is to also check the dependency version if it's
installed.

**What this change does**
Adds function that captures output of `prelude --help` and is
called if `prelude` has been verified to be installed and in the PATH.

**Any side-effects?**
Note that `prelude` displays version information as output from the
`--help` flag, but errors out. Therefore, the version must be parsed
from a general help message that's captured from stderr rather than
stdout. This approach will be invalid if/when FSL fixes this
behaviour.

* Add CLI to dump environment and package info

Resolves #161

**Why this change was necessary**
It's useful to the user to have a CLI tool to dump environment info
into stdout to help ST contributors with debugging.

**What this change does**
Adds 2 helper functions to get environment information including OS
and CPU details, as well as resource usage.

**Any side-effects?**
New dependency is added, so the shimmingtoolbox package will have to
be reinstalled locally.

**Additional context/notes/links**
https://github.com/shimming-toolbox/shimming-toolbox/issues/161#issuecomment-725028653

* Add tests for check_env module

Resolves #161

**Why this change was necessary**
The newly added environment and soft dependency checker features were
missing tests.

**What this change does**
Adds missing tests to verify functionality of check_env module.

**Any side-effects?**
Increased test coverage.

* Add 2 new CLI endpoints for check_env module

Resolves #161

**Why this change was necessary**
CLI endpoints should be accessible with the canonical `st_` prefix.

**What this change does**
Adds CLI endpoints for check_env module functions
`check_dependencies()` and `dump_env_info()`.

**Any side-effects?**
2 new CLI endpoints added to `setup.py`.

* Prettify dependency check output

Resolves https://github.com/shimming-toolbox/shimming-toolbox/pull/176#pullrequestreview-530922591
Related to #161

**Why this change was necessary**
A prettier output is more user-friendly and assists with readability[1].

**What this change does**
Uses snippets from SCT to yield prettier output[2] for dependency checks.

**Additional context/notes/links**
[1] https://github.com/shimming-toolbox/shimming-toolbox/issues/161#issuecomment-725028653
[2] https://github.com/neuropoly/spinalcordtoolbox/blob/master/scripts/sct_check_dependencies.py