CI: Cache SCT if its commit hash is unchanged
Resolves #216

Clones SCT on every run and stores its HEAD commit hash as the cache
key. If this key doesn't match the previously stored cache key, then
SCT is reinstalled and its install location is added to the cache.

The end goal of caching SCT in this workflow is to speed up CI runs -
SCT doesn't need to be reinstalled from scratch if it hasn't changed.