Cache SCT in GitHub Actions CI workflow (#225)
* CI: Cache SCT if its commit hash is unchanged

Resolves #216

Clones SCT on every run and stores its HEAD commit hash as the cache
key. If this key doesn't match the previously stored cache key, then
SCT is reinstalled and its install location is added to the cache.

The end goal of caching SCT in this workflow is to speed up CI runs -
SCT doesn't need to be reinstalled from scratch if it hasn't changed.

Caches are only specified for workflow runner OSes, not Python version.
This keeps us under the 5GB repo cache limit. SCT is assumed to not
change much between Python 3.7 and 3.8, so this is *probably* safe.

Note that the `hashFiles(path)` function can only be passed a `path`
inside the `GITHUB_WORKSPACE` directory.

See https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions#hashfiles