Drop actions/setup-python
'make install' intsalls in conda, so we only actually test or support
python 3.7, and it's conda's python, not `actions/setup-python`'s.

Dropping this cuts the testing time by over half, as the "3.8" runs
were actually the same as the "3.7" runs, and in both cases
`actions/setup-python` was redundant.

If we head in a direction where the code is stable and cross-platform
enough that we can ease out of demanding `conda`, then
`actions/setup-python` could become useful again.