Share a single conda environment. (#37)
pst_venv and st_venv were separated due to version conflicts -- we believe
over wxPython. We have lost the git log explaining exactly what they were.

They seem to be compatible now, especially after https://github.com/shimming-toolbox/fsleyes-plugin-shimming-toolbox/pull/35

so we can recombine everything into a single env.

Notice that this switches from using Miniconda-latest to Miniconda-py3.7, so that the base environment has
the version of python we expect. This is safe because this conda install is *just* for shimming-toolbox.

This corresponds to https://github.com/shimming-toolbox/shimming-toolbox/pull/379 

Co-authored-by: Alexandre D'Astous <po09i@hotmail.com>