Faster B0 shimming by changing some mathematical expressions (#432)
* Change Residuals mse and his jacobian to a faster one
In Python, it's faster to do a matrix multiplication than a np.sum(A*B), so I changed the np.sum in the residuals mse and Jacobian to a matrix multiplication

* Change a np.sum to a matrix multiplication
It's faster on python to do a matrix multiplication than a np.sum. (I checked that the values from both methods are often the same, and from a mathematical pov it should always be the same)

* Declare a self object in the init method

* Change np.sum with A@B

* Fixed an error in the Jacobian
Fixed an error in the Jacobian that changed a little bit the currents in the end. I also changed the mathematical expression of the Jacobian to vectorize it to make it faster. In addition, I changed the formula of residuals mse to an equivalent one, to make it faster

 Please enter the commit message for your changes. Lines starting

* Removed unessary comment

* Fixed useless comments

* Apply suggestions from code review

Fixed comments/ style

Co-authored-by: Alex Dastous <47249340+po09i@users.noreply.github.com>

* Adressed the comment of the PR

* adressed pr comments

* Update sequencer.py

* Reply to comments on PR

* Prevent the calculation of merged_coils if not necessary

* Fixed a comment

* Update shimmingtoolbox/shim/sequencer.py

Co-authored-by: Alex Dastous <47249340+po09i@users.noreply.github.com>

* Update sequencer.py

---------

Co-authored-by: Alex Dastous <47249340+po09i@users.noreply.github.com>